* Compile clean in ubuntu 16.04 * requires package libqglviewer-dev "sudo apt-get install libqglviewer-dev" * renamed some files and includes, probably created under windows, with case conflict.
