+++ GCV-3.6.1-T4.qt52/src/forms/mainwindow.ui
@@ -1246,7 +1246,7 @@
                 <property name="smallDecimalPoint">
                  <bool>false</bool>
                 </property>
-                <property name="numDigits">
+                <property name="digitCount">
                  <number>3</number>
                 </property>
                 <property name="segmentStyle">
