+++ GCV-3.6.1-T4.qt52/src/log4qt/helpers/factory.h
@@ -29,7 +29,7 @@
 /******************************************************************************
  * Dependencies
  ******************************************************************************/
-
+#include <QObject>
 #include <QtCore/QHash>
 #include <QtCore/QMutex>
 #include <QtCore/QStringList>
