+++ GCV-3.6.1-T4.qt52/src/visu3D/viewer3D.cpp
@@ -9,10 +9,10 @@
 
 #include "viewer3D.h"
 #include <math.h>
-#include "point3D.h"
-#include "line3D.h"
-#include "arc3d.h"
-#include "box3D.h"
+#include "Point3D.h"
+#include "Line3D.h"
+#include "Arc3D.h"
+#include "Box3D.h"
 #include "version.h"
 
 // Constructor must call the base class constructor.
